+++ CHANGES.md
@@ -7,6 +7,7 @@ Change Log
   * All `Matrix2`, `Matrix3`, `Matrix4` and `Quaternion` functions that take a `result` parameter now require the parameter (except for functions starting with `from`).
   * Removed the following from the Cesium API: `Transforms.earthOrientationParameters`, `EarthOrientationParameters`, `EarthOrientationParametersSample`, `Transforms.iau2006XysData`, `Iau2006XysData`, `Iau2006XysSample`, `IauOrientationAxes`, `TimeConstants`, `Scene.frameState`, `FrameState`, `EncodedCartesian3`, `EllipsoidalOccluder`, and `FAR`.  These are still available but are not part of the official API and may change in future versions.
   * Removed `DynamicObject.vertexPositions`.  Use `DynamicWall.positions`, `DynamicPolygon.positions`, and `DynamicPolyline.positions` instead.
+  * Removed `Primitive.allow3DOnly`, set the `Scene` level option, `scene3DOnly`, instead.
   * The `DynamicScene` layer has been renamed to `DataSources` additionally, the following objects have all been renamed.
     * `DynamicBillboard` -> `BillboardGraphics`
     * `DynamicBillboardVisualizer` -> `BillboardVisualizer`
@@ -38,10 +39,40 @@ Change Log
   * `DataSource.dynamicObjects` has been renamed to `DataSource.entities`
   * `EntityCollection.getObjects()` and `CompositeEntityCollection.getObjects()` have been made properties and renamed `EntityCollection.entities` and `CompositeEntityCollection.entities`. 
   * `Viewer.trackedObject` and `Viewer.selectedObject` have been renamed to `Viewer.trackedEntity` and `Viewer.selectedEntity` when using the `viewerEntityMixin`.
+  * Renamed functions for naming consistency
+    * `BoundingSphere.getPlaneDistances` -> `BoundingSphere.computePlaneDistances`
+    * `Cartesian[2,3,4].getMaximumComponent` -> `Cartesian[2,3,4].maximumComponent`
+    * `Cartesian[2,3,4].getMinimumComponent` -> `Cartesian[2,3,4].minimumComponent`
+    * `Cartesian[2,3,4].getMaximumByComponent` -> `Cartesian[2,3,4].maximumByComponent`
+    * `Cartesian[2,3,4].getMinimumByComponent` -> `Cartesian[2,3,4].minimumByComponent`
+    * `CubicRealPolynomial.realRoots` -> `CubicRealPolynomial.computeRealRoots`
+    * `CubicRealPolynomial.discriminant` -> `CubicRealPolynomial.computeDiscriminant`
+    * `JulianDate.getTotalDays` -> `JulianDate.totalDyas`
+    * `JulianDate.getSecondsDifference` -> `JulianDate.secondsDifference`
+    * `JulianDate.getDaysDifference` -> `JulianDate.daysDifference`
+    * `JulianDate.getTaiMinusUtc` -> `JulianDate.computeTaiMinusUtc`
+    * `Matrix3.getEigenDecompostion` -> `Matrix3.computeEigenDecomposition`
+    * `Occluder.getVisibility` -> `Occluder.computeVisibility`
+    * `Occluder.getOccludeePoint` -> `Occluder.computerOccludeePoint`
+    * `QuadraticRealPolynomial.discriminant` -> `QuadraticRealPolynomial.computeDiscriminant`
+    * `QuadraticRealPolynomial.realRoots` -> `QuadraticRealPolynomial.computeRealRoots`
+    * `QuarticRealPolynomial.discriminant` -> `QuarticRealPolynomial.computeDiscriminant`
+    * `QuarticRealPolynomial.realRoots` -> `QuarticRealPolynomial.computeRealRoots`
+    * `Quaternion.getAxis` -> `Quaternion.computeAxis`
+    * `Quaternion.getAngle` -> `Quaternion.computeAngle`
+    * `Quaternion.innerQuadrangle` -> `Quaternion.computeInnerQuadrangle`
+    * `Rectangle.getSouthwest` -> `Rectangle.southwest`
+    * `Rectangle.getNorthwest` -> `Rectangle.northwest`
+    * `Rectangle.getSoutheast` -> `Rectangle.southeast`
+    * `Rectangle.getNortheast` -> `Rectangle.northeast`
+    * `Rectangle.getCenter` -> `Rectangle.center`
+    * `CullingVolume.getVisibility` -> `CullingVolume.computeVisibility`
   * Sandcastle examples now automatically wrap the example code in RequireJS boilerplate.  To upgrade any custom examples, copy the code into an existing example (such as Hello World) and save a new file.
 * Added northUpEast transform to help support display of glTF models because Y is their up axis.
 * Cesium can now render an unlimited number of imagery layers, no matter how few texture units are supported by the hardware.
 * Added `Primitive.ready`.
+* Prevent primitives from flashing off and on when modifying static DataSources.
+* Added `scene3DOnly` options to `Viewer`, `CesiumWidget`, and `Scene` constructors. This setting optimizes memory usage and performance for 3D mode at the cost of losing the ability to use 2D or Columbus View.
 
 Beta Releases
 -------------